# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: H Fans API - World App Mini App creator platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /creator/subscribers:
    get:
      operationId: getCreatorSubscribers
      tags: [creators]
      summary: Get creator's subscribers
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Subscribers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriberListResponse"
  /creator/earnings:
    get:
      operationId: getCreatorEarnings
      tags: [creators]
      summary: Get creator earnings
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [week, month, year, all]
            default: month
      responses:
        "200":
          description: Earnings data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EarningsResponse"
  /creator/analytics:
    get:
      operationId: getCreatorAnalytics
      tags: [analytics]
      summary: Get creator analytics
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [week, month, year]
            default: month
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsResponse"
  /creator/vault:
    get:
      operationId: getVault
      tags: [creators]
      summary: Get media vault
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Vault items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultListResponse"
...
  /creator/lists:
    get:
      operationId: getFanLists
      tags: [creators]
      summary: Get fan lists
      responses:
        "200":
          description: Fan lists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FanListResponse"

    post:
      operationId: createFanList
      tags: [creators]
      summary: Create a fan list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFanListRequest"
      responses:
        "201":
          description: Created list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FanList"
  /creator/settings:
    get:
      operationId: getCreatorSettings
      tags: [creators]
      summary: Get creator settings
      responses:
        "200":
          description: Creator settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatorSettings"

    put:
      operationId: updateCreatorSettings
      tags: [creators]
      summary: Update creator settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCreatorSettingsRequest"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatorSettings"
  /creator/goals:
    get:
      operationId: getGoals
      tags: [creators]
      summary: Get fundraising goals
      responses:
        "200":
          description: Goals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalListResponse"

    post:
      operationId: createGoal
      tags: [creators]
      summary: Create a fundraising goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGoalRequest"
      responses:
        "201":
          description: Created goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"

  # SEARCH
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        bannerUrl:
          type: string
        walletAddress:
          type: string
        role:
          type: string
          enum: [fan, creator, admin]
        isWorldIdVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, username, role]
    CreatorProfile:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        bannerUrl:
          type: string
        subscriptionPriceWld:
          type: string
        subscriberCount:
          type: integer
        postCount:
          type: integer
        likeCount:
          type: integer
        isSubscribed:
          type: boolean
        isVerified:
          type: boolean
        isWorldIdVerified:
          type: boolean
        socialLinks:
          $ref: "#/components/schemas/SocialLinks"
      required: [id, username]
    SocialLinks:
      type: object
      properties:
        twitter:
          type: string
        instagram:
          type: string
        tiktok:
          type: string
        spotify:
          type: string
...
    MediaItem:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
        type:
          type: string
          enum: [image, video, audio]
        thumbnailUrl:
          type: string
        blurUrl:
          type: string
        duration:
          type: integer
      required: [id, url, type]
    PollOption:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        voteCount:
          type: integer
      required: [id, text, voteCount]
    Post:
      type: object
      properties:
        id:
          type: string
        creatorId:
          type: string
        creator:
          $ref: "#/components/schemas/CreatorProfile"
        text:
          type: string
        media:
          type: array
          items:
            $ref: "#/components/schemas/MediaItem"
        isPpv:
          type: boolean
        ppvPriceWld:
          type: string
        isUnlocked:
          type: boolean
        isFreeForSubscribers:
          type: boolean
        likeCount:
          type: integer
        commentCount:
          type: integer
        isLiked:
          type: boolean
        isBookmarked:
          type: boolean
        postType:
          type: string
          enum: [text, photo, video, poll, quiz]
        pollOptions:
          type: array
          items:
            $ref: "#/components/schemas/PollOption"
        pollEndsAt:
          type: string
          format: date-time
        scheduledAt:
          type: string
          format: date-time
        goalAmount:
          type: string
        goalCurrent:
          type: string
        goalTitle:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, creatorId, createdAt]
...
    ChartDataPoint:
      type: object
      properties:
        date:
          type: string
        value:
          type: number
        label:
          type: string
      required: [date, value]
    SubscriberListResponse:
      type: object
      properties:
        subscribers:
          type: array
          items:
            $ref: "#/components/schemas/SubscriberInfo"
        total:
          type: integer
        page:
          type: integer
      required: [subscribers, total, page]
    SubscriberInfo:
      type: object
      properties:
        userId:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
        subscribedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        totalSpentWld:
          type: string
        isTopFan:
          type: boolean
      required: [userId]
    EarningsResponse:
      type: object
      properties:
        totalWld:
          type: string
        subscriptionsWld:
          type: string
        ppvWld:
          type: string
        tipsWld:
          type: string
        breakdown:
          type: array
          items:
            $ref: "#/components/schemas/ChartDataPoint"
      required: [totalWld]
    AnalyticsResponse:
      type: object
      properties:
        views:
          type: integer
        uniqueVisitors:
          type: integer
        likes:
          type: integer
        comments:
          type: integer
        newSubscribers:
          type: integer
        revenue:
          type: string
        engagementRate:
          type: number
        viewsChart:
          type: array
          items:
            $ref: "#/components/schemas/ChartDataPoint"
        revenueChart:
          type: array
          items:
            $ref: "#/components/schemas/ChartDataPoint"
        topPosts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
      required: [views, likes, comments, newSubscribers]
    VaultItem:
      type: object
      properties:
        id:
          type: string
        mediaUrl:
          type: string
        thumbnailUrl:
          type: string
        mediaType:
          type: string
          enum: [image, video]
        usedInPostCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, mediaUrl, mediaType, createdAt]
    VaultListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/VaultItem"
        total:
          type: integer
        page:
          type: integer
      required: [items, total, page]
    FanList:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        fanCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, fanCount]
    FanListResponse:
      type: object
      properties:
        lists:
          type: array
          items:
            $ref: "#/components/schemas/FanList"
      required: [lists]
    CreateFanListRequest:
      type: object
      properties:
        name:
          type: string
        fanIds:
          type: array
          items:
            type: string
      required: [name]
    CreatorSettings:
      type: object
      properties:
        subscriptionPriceWld:
          type: string
        welcomeMessageEnabled:
          type: boolean
        welcomeMessageText:
          type: string
        autoRenewEnabled:
          type: boolean
        commentsEnabled:
          type: boolean
        tipsEnabled:
          type: boolean
        watermarkEnabled:
          type: boolean
        platformWalletAddress:
          type: string
      required: [subscriptionPriceWld]
    UpdateCreatorSettingsRequest:
      type: object
      properties:
        subscriptionPriceWld:
          type: string
        welcomeMessageEnabled:
          type: boolean
        welcomeMessageText:
          type: string
        autoRenewEnabled:
          type: boolean
        commentsEnabled:
          type: boolean
        tipsEnabled:
          type: boolean
    Goal:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        targetAmountWld:
          type: string
        currentAmountWld:
          type: string
        isComplete:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, title, targetAmountWld, currentAmountWld]
    GoalListResponse:
      type: object
      properties:
        goals:
          type: array
          items:
            $ref: "#/components/schemas/Goal"
      required: [goals]
    CreateGoalRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        targetAmountWld:
          type: string
        postId:
          type: string
      required: [title, targetAmountWld]
...