# lib/api-zod/src/generated/types/fanList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * H Fans API - World App Mini App creator platform
 * OpenAPI spec version: 0.1.0
 */

export interface FanList {
  id: string;
  name: string;
  fanCount: number;
  createdAt?: Date;
}