# lib/api-zod/src/generated/types/subscriberListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * H Fans API - World App Mini App creator platform
 * OpenAPI spec version: 0.1.0
 */
import type { SubscriberInfo } from "./subscriberInfo";

export interface SubscriberListResponse {
  subscribers: SubscriberInfo[];
  total: number;
  page: number;
}